Add mos -X esp32-gen-key: encr. key generation
Can be used to generate a key in any of the three key slots.

Keys are generated using crypto/rand, quality depends on the host platform.

Typical usage:

```
$ mos -X esp32-gen-key --esp32-enable-flash-encryption flash_enc_key fe_key.bin
$ mos -X esp32-gen-key boot_sign_key bs_key.bin
```

The default is to read and write-protect the key, if for whatever reason that is undesirable, set --esp32-protect-key=false

PUBLISHED_FROM=2847b8bba8cba9772368ac59fa68aeee16ab52fe