Add checksums to RPC frames sent over UART
Thanks to Frozen being permissive, we can enable it in the mos -> device direction today.
device -> mos will have to wait for a while, but that is not that important anyway,
as the primary goal is to be able to detect corruption due to device's FIFO overrun.

PUBLISHED_FROM=7680a3c35bd64cc3ba50f4d4730c799c2d1410a5