Reinstate chunked RPC sending for non-ESP devices
ESP devices have long enough FIFOs for hunking to not be necessary.
So we start conservatively - with chunking, but if we sniff an ESP device on the other end (by checking arch in a response to GetInfo), we set the RPC options to disable chunking.

PUBLISHED_FROM=cca96f3ac0526a2575e9bad7d9f9f0d84ff11326