Don't use user package in mos
The reason is that for some reason it doesn't play nicely with the
static build: sometimes it panics when user.Current() is called, in
other cases it fails more gracefully but still fails.

So this is a workaround.

PUBLISHED_FROM=c45ec7b06c2c849edb9ef102e77f1d4ff2feabb8