Fix compiling of mgos_deps_init()
The `mgos_deps_init()` function thing was placed in the temp dir, but
the dir was not mounted to docker image properly.

But, if some custom config schema is provided in mos.yml as well, then
another file was placed into the temp dir, and this caused it to be
mounted, so `mgos_deps_init()` was compiled as well.

This commit makes tmp dir to be always mounted properly to the docker
image.

PUBLISHED_FROM=07178e67912d816443637ac797bb4a2da3df0858