Fix mos to handle clean libs of non-clean libs
Consider: app depends on lib A, which is dirty, and A depends on lib B,
which is clean so it's not uploaded.

The bug was that lib A was uploaded to the remote builder, but lib B
wasn't taken care of at all.

PUBLISHED_FROM=679a0ddaaee50261ce5f5a0af09f054bf9d2b489