BREAKING: Increase FS to 256K; change flash layout
With this change we increase default FS size to 256K and change the flash layout on ESP8266.
On ESP32 the change is less drastic, since it has partition table.

On ESP8266 it is no longer possible to flash FW build with 4M layout on a smaller device, but it is still possible to select 1M or 2M layout by setting FLASH_SIZE.

However, in both cases the chage is not compatible with OTA, it's not possible to perform OTA from FW before this change to FW after it.
Hopefully this is the last time we do this.

Also, we use this opportunity to enable SPIFFS file metadata on ESP8266.

CC3200 is not affected because it has no space on flash left.

PUBLISHED_FROM=95becfb67814d2f7ddebe1af2ba41baf8db4502a