Implement essentials of digest auth
There is a helper function which can be used in any RPC handler like
this:

```c
  if (!mg_rpc_check_digest_auth(ri, mg_mk_str("my_realm"))) {
    ri = NULL;
    return;
  }
```

And a new config value: `rpc.passwd_file`; if not empty, it should be
the file in htdigest format.

And there is a flag for mos: `--rpc-creds`, whose value can be either
`username:passwd` or `@filename`, where the given filename should
contain `username:passwd`.

So when `mos call` command is issued, and the response is 401 with a
properly structured message, mos will generate digest response, send
back, and return the result to the caller.

Not implemented yet:

- Any sort of auth data persistence; every request of a protected
  resource has to be issued twice;
- Client part on the firmware side is not yet able to react on 401
  properly.

Also, commented some code in `cloud/deployer`, because *clubby.Instance
is incompatible with the ota.Instance, which wants Call() to take the
mgrpc.GetCredsCallback. This deployer is not used anyway.

PUBLISHED_FROM=ab1769dc09a75ec95a73933219606628830e74ff