Support building latest apps with non-latest mos
When cloning a non-latest version of a lib fails, resort to the latest
version of the lib, and print a warning, which depends on whether there
is a newer released mos. If there is:

```
WARNING: the lib "testlibs" does not have version 1.14. Resorting to latest, but the build might fail.
It usually happens if you clone the latest version of some example app, and try to build it with the mos tool which is older than the lib (in this case, "testlibs").
There is a newer version of the mos tool available: 1.15, try to update mos tool (mos update), and build again. Alternatively, you can build the version 1.14 of an app (git checkout 1.14).
```

If not:

```
WARNING: the lib "testlib" does not have version 1.15. Resorting to latest, but the build might fail.
It usually happens if you clone the latest version of some example app, and try to build it with the mos tool which is older than the lib (in this case, "testlib").
Consider using the version 1.15 of the app (git checkout 1.15), or using latest mos tool (mos update latest).
```

If there is a newer released version,

PUBLISHED_FROM=81365bcc72f55039213bf2b7dc32669c56a4a7af