Improve lib sources handling
So now, when we expand libs' manifests, we don't throw all sources in
one pile; instead, we collect them separately for each lib, then we
expand vars, absolutize them and convert to paths to concrete existing
files, and then, depending on --prefer-prebuilt-libs and whether the lib
has existing sources, we add to the main manifest either lib's sources
or its prebuilt binary.

An incidental improvement cesanta/dev#1 is that we don't need a dirty hack when
building libs: for libs we don't need any deps sources to be included in
the resulting .a file, so previously we had to reread the original
manifest and use sources from that original manifest instead of the
expanded one; now we just bypass the libs sources expansion for libs,
which is obviously cleaner.

An incidental improvement cesanta/dev#2 is that all globs are resolved on the mos
side, not in make; so `APP_SOURCES` is set to a list of concrete
existing files.

`mos_final.yml` is now really final; it contains absolute paths to all
exact files, so there are no vars, no globs, etc. Only exact files.

This also means that `eval-manifest-expr` now needs *everything*: not
only libs, but also mongoose-os and all other modules (if any). If their
location was not provided in cli, they will be cloned.

PUBLISHED_FROM=939e6b00102a61ade014666dbccfd6a6883f2a3a