Add support for fetching assets from GitLab
It is currently not as useful as GitHub's due inability to attach binaries to releases
(https://gitlab.com/gitlab-org/gitlab/-/issues/17838)
and not being able to download attachments
(https://gitlab.com/gitlab-org/gitlab/-/issues/24155).

It is possible to store files on a separate server and add links to them to the release metadata.