Bring back lib deduping/caching by name
Make explicit setting of library names stricter, obviating the need to read
manifests of overridden libraries.  Slightly reduce build time as a side effect.

prepareLib(): include an extensive commentary of explicit library naming with
rationales for allowing or forbidding every case of library name setting.
Provide unit tests for all 9 cases commented, plus a test for the obviated
reading of an unreadable but overridden library.

The primary driver for this change is the use case of overriding a library
where the library source being overridden is inaccessible during building.  This
valuable use case can't be supported if determining the library name requires
reading the manifest of the library being overridden.